backend/requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filters==1.3.2
filters-django==1.0.5
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
